change HydraBase quote to not mention master/slave terminology
In order to promote inclusiveness, and actively work against discrimination in tech, several projects have started to evaluate the usage of the terms master/slave to replace them with primary/replica or leader/follower, since master/slave has negative connotations. Projects include:

- Drupal: https://www.drupal.org/node/2275877
- Django: https://github.com/django/django/pull/2692

Extensive discussion has happened in both cases, which I hope reduces the amount of discussion to get this change landed.

Quoting from Wikipedia "Master/slave_(technology)":

Appropriateness of usage

In 2003, the County of Los Angeles in California asked that manufacturers, suppliers and contractors to stop using "master" and "slave" terminology on its products. The controversial decision is taken by the county "based on the cultural diversity and sensitivity of Los Angeles County". Following outcries about the decision, the County of Los Angeles issued a statement saying that the decision was "nothing more than a request". Due to the controversy, the term was selected as the most politically incorrect word in 2004 by Global Language Monitor.

In May 2014, github user fcurella submitted a pull request to the github repository for the Python framework django, initially changing it to leader/follower and finally to primary/replica. This triggered an active discussion of the appropriateness of the master/slave terminology as well as the appropriateness of the change.

In June 2014, Drupal 8 did the same as django did, citing that the word "replica" is already in use by IBM, Microsoft, Engine Yard, Amazon Web Services, and ACM.