add alt tags and meta description tag
The lack of alt tags was hurting the lighthouse score for both accessibility and seo, so that's an easy win. The reason for including `alt=""` for the icon is that if there is no alt tag, some screen readers will read the `src` attribute of the `img` element, which isn't very nice.

The description is also an easy win.